+++ java/src/main/java/org/rocksdb/TickerType.java
@@ -740,6 +740,16 @@ public enum TickerType {
      */
     BLOB_DB_CACHE_BYTES_WRITE((byte) -0x34),
 
+    /**
+     * Number of lookup in prefetch buffer that finds its data
+     */
+    PREFETCH_BUFFER_HIT((byte) -0x3A),
+
+    /**
+     * Number of lookup in prefetch buffer that doesn't find its data
+     */
+    PREFETCH_BUFFER_MISS((byte) -0x3B),
+
     TICKER_ENUM_MAX((byte) 0x5F);
 
     private final byte value;